+++ integration/spark/app/src/main/java/io/openlineage/spark/agent/lifecycle/SparkSQLExecutionContext.java
@@ -42,6 +42,10 @@ import org.apache.spark.sql.execution.ui.SparkListenerSQLExecutionStart;
 class SparkSQLExecutionContext implements ExecutionContext {
 
   private static final String NO_EXECUTION_INFO = "No execution info {}";
+  private static final String SPARK_JOB_TYPE = "JOB";
+  private static final String SPARK_INTEGRATION = "SPARK";
+  private static final String SPARK_PROCESSING_TYPE_BATCH = "BATCH";
+  private static final String SPARK_PROCESSING_TYPE_STREAMING = "STREAMING";
   private final long executionId;
   private final OpenLineageContext olContext;
   private final EventEmitter eventEmitter;
@@ -90,6 +94,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
                 .eventTime(toZonedTime(startEvent.time()))
                 .eventType(eventType),
             buildJob(olContext.getQueryExecution().get()),
+            getJobFacetsBuilder(olContext.getQueryExecution().get()),
             startEvent);
 
     log.debug("Posting event for start {}: {}", executionId, event);
@@ -129,6 +134,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
                 .eventTime(toZonedTime(endEvent.time()))
                 .eventType(eventType),
             buildJob(olContext.getQueryExecution().get()),
+            getJobFacetsBuilder(olContext.getQueryExecution().get()),
             endEvent);
 
     log.debug("Posting event for end {}: {}", executionId, OpenLineageClientUtils.toJson(event));
@@ -155,6 +161,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
                 .eventTime(ZonedDateTime.now(ZoneOffset.UTC))
                 .eventType(RUNNING),
             buildJob(olContext.getQueryExecution().get()),
+            getJobFacetsBuilder(olContext.getQueryExecution().get()),
             stageSubmitted);
 
     log.debug("Posting event for stage submitted {}: {}", executionId, event);
@@ -180,6 +187,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
                 .eventTime(ZonedDateTime.now(ZoneOffset.UTC))
                 .eventType(RUNNING),
             buildJob(olContext.getQueryExecution().get()),
+            getJobFacetsBuilder(olContext.getQueryExecution().get()),
             stageCompleted);
 
     log.debug("Posting event for stage completed {}: {}", executionId, event);
@@ -217,6 +225,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
                 .eventTime(toZonedTime(jobStart.time()))
                 .eventType(eventType),
             buildJob(olContext.getQueryExecution().get()),
+            getJobFacetsBuilder(olContext.getQueryExecution().get()),
             jobStart);
 
     log.debug("Posting event for start {}: {}", executionId, event);
@@ -260,6 +269,7 @@ class SparkSQLExecutionContext implements ExecutionContext {
                 .eventTime(toZonedTime(jobEnd.time()))
                 .eventType(eventType),
             buildJob(olContext.getQueryExecution().get()),
+            getJobFacetsBuilder(olContext.getQueryExecution().get()),
             jobEnd);
 
     log.debug("Posting event for end {}: {}", executionId, event);
@@ -287,6 +297,35 @@ class SparkSQLExecutionContext implements ExecutionContext {
         .newJobBuilder()
         .namespace(this.eventEmitter.getJobNamespace())
         .name(normalizeName(name) + "." + normalizeName(node.nodeName()));
+  }
+
+  /**
+   * Getting the job type facet for Spark jobs. Values: job type: `JOB`, job integration: `SPARK`,
+   * processing type: can be `batch` or `streaming` based on
+   * queryExecution.optimizedPlan().isStreaming()
+   *
+   * @param queryExecution
+   * @return OpenLineage.JobTypeJobFacet
+   */
+  private OpenLineage.JobTypeJobFacet getJobTypeJobFacet(QueryExecution queryExecution) {
+    final String processingType;
+    // Determine processing type
+    if (queryExecution.optimizedPlan().isStreaming()) {
+      processingType = SPARK_PROCESSING_TYPE_STREAMING;
+    } else {
+      processingType = SPARK_PROCESSING_TYPE_BATCH;
+    }
+
+    return openLineage
+        .newJobTypeJobFacetBuilder()
+        .jobType(SPARK_JOB_TYPE)
+        .processingType(processingType)
+        .integration(SPARK_INTEGRATION)
+        .build();
+  }
+
+  private OpenLineage.JobFacetsBuilder getJobFacetsBuilder(QueryExecution queryExecution) {
+    return openLineage.newJobFacetsBuilder().jobType(getJobTypeJobFacet(queryExecution));
   }
 
   // normalizes string, changes CamelCase to snake_case and replaces all non-alphanumerics with '_'